+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Update using ```composer update```.
 * It is now possible to eager load relations on a result set using the `ResultSet::include()` method.
 * Added `ORM::includes()` method that returns `true` if a relation has been loaded and `false` if not.
 * Added `PaginationInterface::toArray()` method.
-* The `Pagination` class now implements the `JsonSerializable` interface.
+* Added `PaginationInterface::toJSON()` method.
+* The `PaginationInterface` interface now extends the `JsonSerializable` interface.
 
 #### Changes